feat: codegen.py seperated into diffrent files
