--- src/codegen.py
@@ -1,2579 +1,0 @@
-from llvmlite import ir, binding
-from typing import TYPE_CHECKING, Dict, Callable, Type
-from astnodes import *
-from lexer import *
-
-
-if TYPE_CHECKING:
-    from compiler import Compiler  # Only for type hints
-
-class SymbolKind(Enum):
-    """Enum for different kinds of symbols."""
-    VARIABLE = auto()
-    FUNCTION = auto()
-    TYPE = auto()
-    PARAMETER = auto()
-    CLASS = auto()
-    UNION = auto()
-    
-from enum import Enum 
-
-class EnumValueType(Enum):
-    """Enum to represent the type of enum values"""
-    INTEGER = "integer"
-    STRING = "string"
-
-class EnumTypeInfo:
-    """Enhanced enum type information that handles both integer and string enums"""
-    
-    def __init__(self, name, enum_value_type, llvm_type, values, ast_node):
-        self.name = name                    # Name of the enum
-        self.enum_value_type = enum_value_type  # EnumValueType.INTEGER or EnumValueType.STRING
-        self.llvm_type = llvm_type         # The underlying LLVM type (i32 for int, i8* for string)
-        self.values = values               # Dictionary mapping member names to their values
-        self.ast_node = ast_node           # Reference to the original AST node
-        
-        # For string enums, we also need to store the LLVM constants
-        self.llvm_constants = {}           # Maps member names to LLVM constants
-    
-    def get_member_value(self, member_name):
-        """Get the value for a specific enum member"""
-        return self.values.get(member_name)
-    
-    def get_llvm_constant(self, member_name):
-        """Get the LLVM constant for a specific enum member"""
-        if self.enum_value_type == EnumValueType.STRING:
-            return self.llvm_constants.get(member_name)
-        else:
-            # For integer enums, create the constant on-the-fly
-            value = self.values.get(member_name)
-            if value is not None:
-                return ir.Constant(self.llvm_type, value)
-        return None
-    
-    def get_llvm_type(self):
-        """Return the LLVM type for this enum"""
-        return self.llvm_type
-    
-    def has_member(self, member_name):
-        """Check if a member exists in this enum"""
-        return member_name in self.values
-    
-    def is_string_enum(self):
-        """Check if this is a string enum"""
-        return self.enum_value_type == EnumValueType.STRING
-    
-    def is_integer_enum(self):
-        """Check if this is an integer enum"""
-        return self.enum_value_type == EnumValueType.INTEGER
-
-class EnumTable:
-    """Centralized enum table to manage all enum types"""
-    
-    def __init__(self):
-        self.enums = {}  # Maps enum names to EnumTypeInfo objects
-    
-    def add_enum(self, name, enum_type_info):
-        """Add an enum to the table"""
-        self.enums[name] = enum_type_info
-    
-    def get_enum(self, name):
-        """Get enum information by name"""
-        return self.enums.get(name)
-    
-    def has_enum(self, name):
-        """Check if an enum exists"""
-        return name in self.enums
-    
-    def remove_enum(self, name):
-        """Remove an enum from the table"""
-        if name in self.enums:
-            del self.enums[name]
-    
-    def clear(self):
-        """Clear all enums"""
-        self.enums.clear()
-
-
-class Symbol:
-    """
-    Represents a symbol in the symbol table.
-    A symbol can be a variable, function, type, etc.
-    """
-    def __init__(self, 
-                 name: str, 
-                 kind: SymbolKind, 
-                 ast_node: Any, 
-                 data_type: Any,
-                 llvm_type: Any = None, 
-                 llvm_value: Optional[ir.Value] = None, 
-                 scope_level: int = 0,
-                 pointer_level: int = 0):
-        self.name = name
-        self.kind = kind
-        self.ast_node = ast_node
-        self.data_type = data_type  # The language type (e.g., 'U8', 'MyStruct')
-        self.llvm_type = llvm_type  # The LLVM type
-        self.llvm_value = llvm_value  # LLVM value or pointer
-        self.scope_level = scope_level
-        self.pointer_level = pointer_level  # 0 = not a pointer, 1 = pointer, 2 = double pointer, etc.
-        # Additional data for specific symbol kinds
-        self.extra_data: Dict[str, Any] = {}
-
-    @property
-    def is_pointer(self) -> bool:
-        """Backward compatibility property."""
-        return self.pointer_level > 0
-
-    def __repr__(self) -> str:
-        pointer_str = f", ptr_level={self.pointer_level}" if self.pointer_level > 0 else ""
-        return f"Symbol(name='{self.name}', kind={self.kind}, type={self.data_type}, scope={self.scope_level}{pointer_str})"
-
-
-class Scope:
-    """Represents a single scope level in the symbol table."""
-    def __init__(self, level: int):
-        self.level = level
-        self.symbols: Dict[str, Symbol] = {}
-
-    def define(self, symbol: Symbol) -> None:
-        """Define a symbol in this scope."""
-        if symbol.name in self.symbols:
-            raise ValueError(f"Symbol '{symbol.name}' already defined in scope {self.level}")
-        self.symbols[symbol.name] = symbol
-
-    def lookup(self, name: str) -> Optional[Symbol]:
-        """Look up a symbol in this scope."""
-        return self.symbols.get(name)
-
-    def get_all_symbols(self) -> List[Symbol]:
-        """Get all symbols in this scope."""
-        return list(self.symbols.values())
-
-
-class SymbolTable:
-    """
-    Symbol table with scope management.
-    Manages multiple scopes for block-level variable declarations.
-    """
-    def __init__(self):
-        self.scopes: List[Scope] = [Scope(0)]  # Start with global scope (level 0)
-        self.current_scope_level = 0
-
-    @property
-    def current_scope(self) -> Scope:
-        """Get the current scope."""
-        return self.scopes[self.current_scope_level]
-
-    def enter_scope(self) -> int:
-        """
-        Enter a new scope level.
-        Returns the new scope level.
-        """
-        self.current_scope_level += 1
-        # Create the new scope if it doesn't exist
-        if len(self.scopes) <= self.current_scope_level:
-            self.scopes.append(Scope(self.current_scope_level))
-        return self.current_scope_level
-
-    def exit_scope(self) -> int:
-        """
-        Exit the current scope level.
-        Returns the new (previous) scope level.
-        """
-        if self.current_scope_level > 0:
-            self.current_scope_level -= 1
-        return self.current_scope_level
-
-    def define(self, symbol: Symbol) -> Symbol:
-        """
-        Define a symbol in the current scope.
-        Sets the scope level on the symbol and adds it to the current scope.
-        """
-        # Set the scope level on the symbol
-        symbol.scope_level = self.current_scope_level
-        # Add to current scope
-        self.current_scope.define(symbol)
-        return symbol
-
-    def lookup(self, name: str, current_scope_only: bool = False) -> Optional[Symbol]:
-        """
-        Look up a symbol by name.
-        If current_scope_only is True, only look in the current scope.
-        Otherwise, look in all scopes from current to global.
-        """
-        if current_scope_only:
-            return self.current_scope.lookup(name)
-        
-        # Search from current scope up to global scope
-        for level in range(self.current_scope_level, -1, -1):
-            symbol = self.scopes[level].lookup(name)
-            if symbol:
-                return symbol
-        
-        return None
-
-    def lookup_in_scope(self, name: str, scope_level: int) -> Optional[Symbol]:
-        """Look up a symbol in a specific scope level."""
-        if 0 <= scope_level < len(self.scopes):
-            return self.scopes[scope_level].lookup(name)
-        return None
-
-    def remove(self, name: str, scope_level: Optional[int] = None) -> bool:
-        """
-        Remove a symbol from the specified scope or current scope if not specified.
-        Returns True if the symbol was found and removed, False otherwise.
-        """
-        if scope_level is None:
-            scope_level = self.current_scope_level
-        
-        if 0 <= scope_level < len(self.scopes):
-            scope = self.scopes[scope_level]
-            if name in scope.symbols:
-                del scope.symbols[name]
-                return True
-        
-        return False
-
-    def exists(self, name: str, current_scope_only: bool = False) -> bool:
-        """Check if a symbol exists in the symbol table."""
-        return self.lookup(name, current_scope_only) is not None
-
-    def get_all_symbols(self, include_parent_scopes: bool = False) -> List[Symbol]:
-        """
-        Get all symbols in the current scope.
-        If include_parent_scopes is True, include symbols from all parent scopes.
-        """
-        if not include_parent_scopes:
-            return self.current_scope.get_all_symbols()
-        
-        all_symbols = []
-        for level in range(self.current_scope_level + 1):
-            all_symbols.extend(self.scopes[level].get_all_symbols())
-        
-        return all_symbols
-
-    def get_functions(self) -> List[Symbol]:
-        """Get all function symbols."""
-        functions = []
-        for scope in self.scopes:
-            for symbol in scope.symbols.values():
-                if symbol.kind == SymbolKind.FUNCTION:
-                    functions.append(symbol)
-        return functions
-
-    def get_types(self) -> List[Symbol]:
-        """Get all type symbols (classes, unions, etc.)."""
-        types = []
-        for scope in self.scopes:
-            for symbol in scope.symbols.values():
-                if symbol.kind == SymbolKind.TYPE or symbol.kind == SymbolKind.CLASS or symbol.kind == SymbolKind.UNION:
-                    types.append(symbol)
-        return types
-
-    def dump(self) -> str:
-        """Dump the symbol table as a string for debugging."""
-        result = "Symbol Table:\n"
-        for level, scope in enumerate(self.scopes):
-            result += f"  Scope level {level}:\n"
-            for name, symbol in scope.symbols.items():
-                result += f"    {symbol}\n"
-        return result
-
-    def __getitem__(self, name: str) -> Optional[Symbol]:
-        """Allow dict-like access to the symbol table."""
-        return self.lookup(name)
-
-    def __contains__(self, name: str) -> bool:
-        """Allow 'in' operator to check if a symbol exists."""
-        return self.exists(name)
-
-
-# Helper functions for creating common types of symbols
-def create_variable_symbol(name: str, ast_node: Any, data_type: Any, 
-                          llvm_type: Any = None, llvm_value: Optional[ir.Value] = None, 
-                          scope_level: int = 0, pointer_level: int = 0) -> Symbol:
-    """Create a variable symbol."""
-    return Symbol(name, SymbolKind.VARIABLE, ast_node, data_type, 
-                 llvm_type, llvm_value, scope_level, pointer_level)
-
-
-def create_function_symbol(name: str, ast_node: Any, return_type: Any, 
-                          parameter_types: List[Any], llvm_function: Optional[ir.Function] = None, 
-                          scope_level: int = 0) -> Symbol:
-    """Create a function symbol."""
-    symbol = Symbol(name, SymbolKind.FUNCTION, ast_node, return_type, 
-                   None, llvm_function, scope_level, 0)  # Functions are not pointers by default
-    symbol.extra_data['parameter_types'] = parameter_types
-    return symbol
-
-
-def create_type_symbol(name: str, ast_node: Any, llvm_type: Any = None, 
-                      scope_level: int = 0) -> Symbol:
-    """Create a type symbol (class, struct, enum, etc.)."""
-    symbol = Symbol(name, SymbolKind.TYPE, ast_node, name, llvm_type, None, scope_level, 0)
-    return symbol
-
-
-# Utility functions for pointer level handling
-def count_pointer_level(type_str: str) -> tuple[str, int]:
-    """
-    Count the pointer level from a type string.
-    Returns (base_type, pointer_level)
-    
-    Examples:
-    - "int" -> ("int", 0)
-    - "int*" -> ("int", 1)
-    - "char**" -> ("char", 2)
-    - "MyStruct***" -> ("MyStruct", 3)
-    """
-    base_type = type_str.rstrip('*')
-    pointer_level = len(type_str) - len(base_type)
-    return base_type, pointer_level
-
-
-def apply_pointer_level(base_llvm_type: Any, pointer_level: int) -> Any:
-    """
-    Apply pointer level to an LLVM base type.
-    
-    Args:
-        base_llvm_type: Base LLVM type (e.g., ir.IntType(32))
-        pointer_level: Number of pointer levels (0 = no pointer, 1 = pointer, etc.)
-    
-    Returns:
-        LLVM type with appropriate pointer levels applied
-    """
-    result_type = base_llvm_type
-    for _ in range(pointer_level):
-        result_type = ir.PointerType(result_type)
-    return result_type
-
-class Codegen:
-    def __init__(self, compiler: "Compiler"):
-        # Initialize the new symbol table
-        self.symbol_table = SymbolTable()
-        
-        self.compiler = compiler
-        self.astnodes = compiler.astnodes
-        if self.compiler.triple:
-            self.triple = self.compiler.target.get_llvm_triple()
-        else:
-            self.triple = ""
-
-        self.node_index = 0
-        self.current_node = self.astnodes[self.node_index]
-        
-        # For storing function and struct information
-        self.function_map = {}
-        self.struct_table = {}
-
-        self.string_literals = []
-
-        # Node type to handler mapping
-        self.node_handlers: Dict[Type, Callable] = {
-            ASTNode.ExpressionNode: self.handle_expression,
-            ASTNode.Block: self.handle_block,
-            ASTNode.VariableDeclaration: self.handle_variable_declaration,
-            ASTNode.VariableAssignment: self.handle_variable_assignment,
-            ASTNode.Return: self.handle_return,
-            ASTNode.FunctionDefinition: self.handle_function_definition,
-            ASTNode.IfStatement: self.handle_if_statement,
-            ASTNode.WhileLoop: self.handle_while_loop,
-            ASTNode.ForLoop: self.handle_for_loop,
-            ASTNode.Comment: self.handle_comment,
-            ASTNode.FunctionCall: self.handle_function_call,
-            ASTNode.Class: self.handle_class,
-            ASTNode.Union: self.handle_union,
-            ASTNode.Break: self.handle_break,
-            ASTNode.Continue: self.handle_continue,
-            ASTNode.VariableIncrement: self.handle_variable_increment,
-            ASTNode.VariableDecrement: self.handle_variable_decrement,
-            ASTNode.Extern: self.handle_extern,
-            ASTNode.CompoundVariableDeclaration: self.handle_compound_variable_declaration,
-            ASTNode.CompoundVariableAssigment: self.handle_compound_variable_assignment,
-            ASTNode.Enum: self.handle_enum,
-            
-        }
-
-        # Define correct LLVM types with appropriate signedness
-        # Boolean is represented as i1
-        bool_type = ir.IntType(1)
-        # Unsigned types
-        u8_type = ir.IntType(8)
-        u16_type = ir.IntType(16)
-        u32_type = ir.IntType(32)
-        u64_type = ir.IntType(64)
-        # Signed types - in LLVM IR, the types are the same but operations differ
-        i8_type = ir.IntType(8)
-        i16_type = ir.IntType(16)
-        i32_type = ir.IntType(32)
-        i64_type = ir.IntType(64)
-        # Other types
-        void_type = ir.VoidType()
-        f32_type = ir.FloatType()
-        f64_type = ir.DoubleType()
-
-        self.type_map = {
-            Datatypes.BOOL: bool_type,
-            Datatypes.U8: u8_type,
-            Datatypes.U16: u16_type,
-            Datatypes.U32: u32_type,
-            Datatypes.U64: u64_type,
-            Datatypes.I8: i8_type,
-            Datatypes.I16: i16_type,
-            Datatypes.I32: i32_type,
-            Datatypes.I64: i64_type,
-            Datatypes.U0: void_type,
-            Datatypes.F32: f32_type,
-            Datatypes.F64: f64_type
-        }
-
-        self.type_signedness = {
-            self.type_map[Datatypes.I8]: True,
-            self.type_map[Datatypes.I16]: True,
-            self.type_map[Datatypes.I32]: True,
-            self.type_map[Datatypes.I64]: True,
-            self.type_map[Datatypes.U8]: False,
-            self.type_map[Datatypes.U16]: False,
-            self.type_map[Datatypes.U32]: False,
-            self.type_map[Datatypes.U64]: False,
-            self.type_map[Datatypes.BOOL]: False,
-        }
-
-        self.signed_int_types = {i8_type, i16_type, i32_type, i64_type}
-        self.unsigned_int_types = {bool_type, u8_type, u16_type, u32_type, u64_type}
-        self.float_types = {f32_type, f64_type}
-
-    def generation_error(self, message: str, node: 'ASTNode'):
-        """Report an error with a formatted message, based on the ASTNode (instead of token)."""
-        
-        # Retrieve line and column information from the ASTNode (assuming these attributes exist)
-        if hasattr(node, 'line') and hasattr(node, 'column'):
-            line = node.line
-            column = node.column
-        else:
-            line = column = -1  # If line/column info is missing, set to -1 for safety
-
-        # If source code is available, try to get the line of code where the error occurred
-        try:
-            error_line = self.compiler.code.splitlines()[line - 1]  # Subtract 1 for 0-based index
-        except IndexError:
-            error_line = "[ERROR: Line out of range]"
-
-        # Align the caret with the column position (ensure it's within bounds)
-        caret_position = " " * (min(column, len(error_line))) + "^"
-
-        # Print error message and source line context
-        print(f"Generation Error: {message} at line {line}, column {column}")
-        print(f"{error_line}")
-        print(f"{caret_position}")
-        
-        # Print detailed node information
-        print(f"Caused by ASTNode: {repr(node)}")
-
-        # Raise an exception with the full error message
-        raise Exception(f"{message} at line {line}, column {column}\n"
-                        f"{error_line}\n"
-                        f"{caret_position}\n"
-                        f"Caused by ASTNode: {repr(node)}")
-
-        
-    def add_function(self, function: ASTNode.FunctionCall):
-        self.function_map[function.name] = function
-        
-    def lookup_function(self, name: str) -> Optional[ASTNode.FunctionDefinition]:
-        """
-        Look up a function by name and return the FunctionDefinition.
-        Returns None if the function is not found.
-        """
-        return self.function_map.get(name)
-
-    def current_node(self):
-        self.current_node = self.astnodes[self.node_index]
-        return self.current_node
-
-    def next_node(self):
-        self.node_index += 1
-        self.current_node = self.astnodes[self.node_index]
-        return self.current_node
-
-
-    def get_llvm_type(self, type: str):
-        if type in self.type_map:
-            return self.type_map[type]
-        # handle pointer types
-        elif type.endswith('*'):
-            base_type = self.get_llvm_type(type[:-1]) # delete the * symbol and get it's type
-            return ir.PointerType(base_type) 
-        else: # other types (classes)
-            # TODO! implement
-            print("Non-Primitive types did not implemented. Returning a generic type (U8*)")
-            return ir.PointerType(ir.IntType(8)) # return generic u8 pointer
-
-    def gen(self):
-        # Initialize LLVM
-        binding.initialize()
-        binding.initialize_native_target()
-        binding.initialize_native_asmprinter()
-        
-        try:
-            # Try to use the specified target
-            print(self.triple)
-            target = binding.Target.from_triple(self.triple)
-        except RuntimeError:
-            # get native target if specified target doesn't available
-            print(f"Warning: Target '{self.triple}' not available, using native target instead")
-            target = binding.Target.from_default_triple()
-        
-        target_machine = target.create_target_machine()
-        
-        # data layout string
-        data_layout = target_machine.target_data
-        
-        # create the module 
-        module = ir.Module(name=self.compiler.file)
-        
-        # Use the actual triple from the target machine to ensure compatibility
-        module.triple = target_machine.triple
-        module.data_layout = str(data_layout)
-        
-        # Build context for code generation
-        self.module = module
-        self.context = ir.context.Context()
-        self.builder = None
-        self.function = None
-        
-        # iterate astnodes for handler 
-        for node in self.astnodes:
-            self.process_node(node)
-            
-        return module
-    
-    def process_node(self, node, **kwargs):
-        # Get the node's class type
-        node_class = type(node)
-        
-        # Look up and call the appropriate handler
-        if node_class in self.node_handlers:
-            return self.node_handlers[node_class](node, **kwargs)
-        else:
-            print(f"Warning: No handler for node type {node_class.__name__}")
-            return None
-    
-    def turn_variable_type_to_llvm_type(self, type: Datatypes):
-        llvm_type  = self.type_map[type]
-        return llvm_type
-
-    def handle_function_definition(self, node: ASTNode.FunctionDefinition, builder: Optional[ir.IRBuilder] = None, **kwargs):
-        """Handle function definition with the new symbol table."""
-        name = node.name
-        return_type = Datatypes.to_llvm_type(node.return_type)
-
-        node_params: List[ASTNode.VariableDeclaration] = node.parameters
-        llvm_params = []
-        param_types = []
-
-        # Parse args
-        for param in node_params:
-            if param.is_user_typed:
-                print("NOT IMPLEMENTED! user typed parameters")
-            if param.is_pointer:
-                print("NOT IMPLEMENTED, function pointer types")
-            
-            param_type = self.type_map[param.var_type]
-            llvm_params.append(param_type)
-            param_types.append(param.var_type)
-        
-        # Create the function type and function
-        func_type = ir.FunctionType(return_type, llvm_params)
-        func = ir.Function(self.module, func_type, name)
-        
-        # Create and store the function symbol
-        func_symbol = create_function_symbol(
-            name=name,
-            ast_node=node,
-            return_type=node.return_type,
-            parameter_types=param_types,
-            llvm_function=func
-        )
-        self.symbol_table.define(func_symbol)
-        
-        # Also store in function map for backward compatibility
-        self.function_map[name] = func
-        
-        # Handle function body if we have one
-        if node.body:
-            # Create a new scope for the function body
-            self.symbol_table.enter_scope()
-            
-            # Create the entry block and builder
-            entry_block = func.append_basic_block("entry")
-            local_builder = ir.IRBuilder(entry_block)
-            
-            # Define function parameters in the symbol table
-            for i, (param, llvm_param) in enumerate(zip(node_params, func.args)):
-                # Allocate space for the parameter
-                param_ptr = local_builder.alloca(llvm_param.type, name=f"{param.name}_param")
-                local_builder.store(llvm_param, param_ptr)
-                
-                # Add parameter to symbol table
-                param_symbol = Symbol(
-                    name=param.name,
-                    kind=SymbolKind.PARAMETER,
-                    ast_node=param,
-                    data_type=param.var_type,
-                    llvm_type=llvm_param.type,
-                    llvm_value=param_ptr,
-                    scope_level=self.symbol_table.current_scope_level
-                )
-                self.symbol_table.define(param_symbol)
-            
-            # Process the function body
-            self.process_node(node.body, builder=local_builder)
-            
-            # Ensure the function has a return statement if needed
-            last_block = local_builder.block
-            if not last_block.is_terminated:
-                if return_type == ir.VoidType():
-                    local_builder.ret_void()
-                else:
-                    # For non-void functions, add a default return value
-                    local_builder.ret(ir.Constant(return_type, 0))
-            
-            # Exit the function scope
-            self.symbol_table.exit_scope()
-        
-        return func
-
-
-    def handle_binary_expression(self, node: ASTNode.ExpressionNode, builder: ir.IRBuilder, var_type, **kwargs):
-        is_debug = self.compiler.debug
-        # Parse the operator
-        operator = node.op
-
-        # Determine type characteristics
-        is_signed = False
-        is_float = False
-        is_integer = False
-        
-        # Try to get type information in different ways
-        if hasattr(node, "var_type") and node.var_type:
-            is_signed = Datatypes.is_signed_type(node.var_type)
-            is_float = Datatypes.is_float_type(node.var_type)
-            is_integer = Datatypes.is_integer_type(node.var_type)
-            if is_debug:
-                print(f"DEBUG - From node var_type: is_signed={is_signed}, is_float={is_float}, is_integer={is_integer}, type={node.var_type}")
-        elif hasattr(var_type, "datatype_name") and var_type.datatype_name:
-            is_signed = Datatypes.is_signed_type(var_type.datatype_name)
-            is_float = Datatypes.is_float_type(var_type.datatype_name)
-            is_integer = Datatypes.is_integer_type(var_type.datatype_name)
-            if is_debug:
-                print(f"DEBUG - From var_type datatype_name: is_signed={is_signed}, is_float={is_float}, is_integer={is_integer}, type={var_type.datatype_name}")
-        else:
-            # Default values based on LLVM type if we can't determine from name
-            if isinstance(var_type, ir.IntType):
-                # Check if this type is in the signed types list
-                signed_types = [
-                    self.type_map[t] for t in [Datatypes.I8, Datatypes.I16, Datatypes.I32, Datatypes.I64]
-                ]
-                is_signed = self.type_signedness.get(var_type, False)
-                is_integer = True
-                if is_debug:
-                    print(f"DEBUG - From LLVM IntType: is_signed={is_signed}, width={var_type.width}, type={var_type}")
-            elif isinstance(var_type, (ir.FloatType, ir.DoubleType)):
-                is_float = True
-                is_integer = False
-                if is_debug:
-                    print(f"DEBUG - From LLVM FloatType: is_float={is_float}, type={var_type}")
-            else:
-                # If we can't determine, default to unsigned integer
-                is_signed = False
-                is_integer = True
-                is_float = False
-                if is_debug:
-                    print(f"DEBUG - Using defaults: is_signed={is_signed}, is_float={is_float}, is_integer={is_integer}, type={var_type}")
-
-        # evaluate left and right expressions
-        left = self.handle_expression(node.left, builder, var_type)
-        right = self.handle_expression(node.right, builder, var_type)
-        
-        # Debug the operation
-        if is_debug:
-            print(f"DEBUG - Operation: {operator}, Left type: {left.type}, Right type: {right.type}")
-        
-        # Make sure both operands have the same type
-        if left.type != right.type:
-            if is_debug:
-                print(f"DEBUG - Type mismatch: converting right operand from {right.type} to {left.type}")
-            
-            # For boolean to integer conversions (i1 to i8, etc.)
-            if right.type.width < left.type.width:
-                if right.type.width == 1:  # Converting from boolean (i1)
-                    right = builder.zext(right, left.type, name="bool_to_int")
-                else:
-                    # Handle other integer size conversions
-                    if is_signed:
-                        right = builder.sext(right, left.type, name="sext")
-                    else:
-                        right = builder.zext(right, left.type, name="zext")
-            elif left.type.width < right.type.width:
-                if left.type.width == 1:  # Converting from boolean (i1)
-                    left = builder.zext(left, right.type, name="bool_to_int")
-                else:
-                    # Handle other integer size conversions
-                    if is_signed:
-                        left = builder.sext(left, right.type, name="sext")
-                    else:
-                        left = builder.zext(left, right.type, name="zext") 
-                
-            # Check if we need to handle float conversions
-            if isinstance(left.type, ir.IntType) and isinstance(right.type, (ir.FloatType, ir.DoubleType)):
-                if is_signed:
-                    left = builder.sitofp(left, right.type, name="int_to_float")
-                else:
-                    left = builder.uitofp(left, right.type, name="uint_to_float")
-            elif isinstance(right.type, ir.IntType) and isinstance(left.type, (ir.FloatType, ir.DoubleType)):
-                if is_signed:
-                    right = builder.sitofp(right, left.type, name="int_to_float")
-                else:
-                    right = builder.uitofp(right, left.type, name="uint_to_float")
-        
-        # After conversion, re-check what types we're working with
-        is_float = isinstance(left.type, (ir.FloatType, ir.DoubleType))
-        is_integer = isinstance(left.type, ir.IntType)
-        
-        # Handle operations based on operator type
-        if operator == operators["ADD"]:
-            return builder.add(left, right, name="sum")
-        elif operator == operators["SUBTRACT"]:
-            return builder.sub(left, right, name="sub")
-        elif operator == operators["MULTIPLY"]:
-            return builder.mul(left, right, name="mul")
-        elif operator == operators["DIVIDE"]:
-            # For integer division
-            if is_integer:
-                if is_signed:
-                    if is_debug:
-                        print(f"DEBUG - Using SIGNED integer division (sdiv)")
-                    return builder.sdiv(left, right, name="sdiv")
-                else:
-                    if is_debug:
-                        print(f"DEBUG - Using UNSIGNED integer division (udiv)")
-                    return builder.udiv(left, right, name="udiv")
-            # For floating point division
-            else:
-                if is_debug:
-                    print(f"DEBUG - Using floating point division (fdiv)")
-                return builder.fdiv(left, right, name="fdiv")
-        elif operator == operators["MODULO"]:
-            # For integer modulo
-            if is_integer:
-                if is_signed:
-                    if is_debug:
-                        print(f"DEBUG - Using SIGNED integer remainder (srem)")
-                    return builder.srem(left, right, name="srem")
-                else:
-                    if is_debug:
-                        print(f"DEBUG - Using UNSIGNED integer remainder (urem)")
-                    return builder.urem(left, right, name="urem")
-            # For floating point modulo
-            else:
-                if is_debug:
-                    print(f"DEBUG - Using floating point remainder (frem)")
-                return builder.frem(left, right, name="frem")
-        elif operator == operators["BITWISE_AND"]:
-            return builder.and_(left, right, name="and")
-        elif operator == operators["BITWISE_OR"]:
-            return builder.or_(left, right, name="or")
-        elif operator == operators["BITWISE_XOR"]:
-            return builder.xor(left, right, name="xor")
-        elif operator == operators["SHIFT_LEFT"]:
-            return builder.shl(left, right, name="shl")
-        elif operator == operators["SHIFT_RIGHT"]:
-            # Arithmetic shift for signed types
-            if is_signed:
-                if is_debug:
-                    print(f"DEBUG - Using arithmetic right shift (ashr) for signed type")
-                return builder.ashr(left, right, name="ashr")
-            # Logical shift for unsigned types
-            else:
-                if is_debug:
-                    print(f"DEBUG - Using logical right shift (lshr) for unsigned type")
-                return builder.lshr(left, right, name="lshr")
-        # Comparison operators
-        elif operator == operators["EQUAL"]:
-            # Ensure operands have the same type for comparison
-            if left.type != right.type:
-                if is_debug:
-                    print(f"DEBUG - Type mismatch in comparison: converting operands to match")
-                if left.type.width > right.type.width:
-                    right = builder.zext(right, left.type, name="zext_for_cmp") if right.type.width == 1 else right
-                else:
-                    left = builder.zext(left, right.type, name="zext_for_cmp") if left.type.width == 1 else left
-            
-            if is_integer:
-                if is_signed:
-                    if is_debug:
-                        print(f"DEBUG - Using SIGNED integer comparison (==)")
-                    return builder.icmp_signed('==', left, right, name="seq")  # Will return i1
-                else:
-                    if is_debug:
-                        print(f"DEBUG - Using UNSIGNED integer comparison (==)")
-                    return builder.icmp_unsigned('==', left, right, name="ueq")  # Will return i1
-            else:
-                if is_debug:
-                    print(f"DEBUG - Using floating point comparison (==)")
-                return builder.fcmp_ordered('==', left, right, name="feq")  # Will return i1
-        elif operator == operators["NOT_EQUAL"]:
-            if is_integer:
-                if is_signed:
-                    if is_debug:
-                        print(f"DEBUG - Using SIGNED integer comparison (!=)")
-                    return builder.icmp_signed('!=', left, right, name="sne") 
-                else:
-                    if is_debug:
-                        print(f"DEBUG - Using UNSIGNED integer comparison (!=)")
-                    return builder.icmp_unsigned('!=', left, right, name="une")
-            else:
-                if is_debug:
-                    print(f"DEBUG - Using floating point comparison (!=)")
-                return builder.fcmp_ordered('!=', left, right, name="fne")
-        elif operator == operators["LESS_THAN"]:
-            if is_integer:
-                if is_signed:
-                    if is_debug:
-                        print(f"DEBUG - Using SIGNED integer comparison (<)")
-                    return builder.icmp_signed('<', left, right, name="slt")
-                else:
-                    if is_debug:
-                        print(f"DEBUG - Using UNSIGNED integer comparison (<)")
-                    return builder.icmp_unsigned('<', left, right, name="ult")
-            else:
-                if is_debug:
-                    print(f"DEBUG - Using floating point comparison (<)")
-                return builder.fcmp_ordered('<', left, right, name="flt")
-        elif operator == operators["LESS_OR_EQUAL"]:
-            if is_integer:
-                if is_signed:
-                    if is_debug:
-                        print(f"DEBUG - Using SIGNED integer comparison (<=)")
-                    return builder.icmp_signed('<=', left, right, name="sle")
-                else:
-                    if is_debug:
-                        print(f"DEBUG - Using UNSIGNED integer comparison (<=)")
-                    return builder.icmp_unsigned('<=', left, right, name="ule")
-            else:
-                if is_debug:
-                    print(f"DEBUG - Using floating point comparison (<=)")
-                return builder.fcmp_ordered('<=', left, right, name="fle")
-        elif operator == operators["GREATER_THAN"]:
-            if is_integer:
-                if is_signed:
-                    if is_debug:
-                        print(f"DEBUG - Using SIGNED integer comparison (>)")
-                    return builder.icmp_signed('>', left, right, name="sgt")
-                else:
-                    if is_debug:
-                        print(f"DEBUG - Using UNSIGNED integer comparison (>)")
-                    return builder.icmp_unsigned('>', left, right, name="ugt")
-            else:
-                if is_debug:
-                    print(f"DEBUG - Using floating point comparison (>)")
-                return builder.fcmp_ordered('>', left, right, name="fgt")
-        elif operator == operators["GREATER_OR_EQUAL"]:
-            if is_integer:
-                if is_signed:
-                    if is_debug:
-                        print(f"DEBUG - Using SIGNED integer comparison (>=)")
-                    return builder.icmp_signed('>=', left, right, name="sge")
-                else:
-                    if is_debug:
-                        print(f"DEBUG - Using UNSIGNED integer comparison (>=)") 
-                    return builder.icmp_unsigned('>=', left, right, name="uge")
-            else:
-                if is_debug:
-                    print(f"DEBUG - Using floating point comparison (>=)")
-                return builder.fcmp_ordered('>=', left, right, name="fge")
-        elif operator == operators["LOGICAL_AND"]:
-            # Perform boolean conversion if needed
-            if left.type.width > 1:
-                if is_signed:
-                    if is_debug:
-                        print(f"DEBUG - Converting SIGNED integer to boolean for LOGICAL_AND")
-                    left_bool = builder.icmp_signed('!=', left, ir.Constant(left.type, 0), name="tobool_left")
-                else:
-                    if is_debug:
-                        print(f"DEBUG - Converting UNSIGNED integer to boolean for LOGICAL_AND")
-                    left_bool = builder.icmp_unsigned('!=', left, ir.Constant(left.type, 0), name="tobool_left")
-            else:
-                left_bool = left
-                
-            if right.type.width > 1:
-                if is_signed:
-                    if is_debug:
-                        print(f"DEBUG - Converting SIGNED integer to boolean for LOGICAL_AND")
-                    right_bool = builder.icmp_signed('!=', right, ir.Constant(right.type, 0), name="tobool_right")
-                else:
-                    if is_debug:
-                        print(f"DEBUG - Converting UNSIGNED integer to boolean for LOGICAL_AND")
-                    right_bool = builder.icmp_unsigned('!=', right, ir.Constant(right.type, 0), name="tobool_right")
-            else:
-                right_bool = right
-                
-            return builder.and_(left_bool, right_bool, name="land")
-        elif operator == operators["LOGICAL_OR"]:
-            # Perform boolean conversion if needed
-            if left.type.width > 1:
-                if is_signed:
-                    if is_debug:
-                        print(f"DEBUG - Converting SIGNED integer to boolean for LOGICAL_OR")
-                    left_bool = builder.icmp_signed('!=', left, ir.Constant(left.type, 0), name="tobool_left")
-                else:
-                    if is_debug:
-                        print(f"DEBUG - Converting UNSIGNED integer to boolean for LOGICAL_OR")
-                    left_bool = builder.icmp_unsigned('!=', left, ir.Constant(left.type, 0), name="tobool_left")
-            else:
-                left_bool = left
-                
-            if right.type.width > 1:
-                if is_signed:
-                    if is_debug:
-                        print(f"DEBUG - Converting SIGNED integer to boolean for LOGICAL_OR")
-                    right_bool = builder.icmp_signed('!=', right, ir.Constant(right.type, 0), name="tobool_right")
-                else:
-                    if is_debug:
-                        print(f"DEBUG - Converting UNSIGNED integer to boolean for LOGICAL_OR")
-                    right_bool = builder.icmp_unsigned('!=', right, ir.Constant(right.type, 0), name="tobool_right")
-            else:
-                right_bool = right
-                
-            return builder.or_(left_bool, right_bool, name="lor")
-        else:
-            raise ValueError(f"Unsupported binary operator: {operator}")
-        
-    def get_variable_pointer(self, name):
-        """Get the LLVM value pointer for a variable."""
-        symbol = self.symbol_table.lookup(name)
-        if not symbol:
-            raise Exception(f"Undefined variable: {name}")
-        
-        return symbol.llvm_value
-
-
-    def handle_primary_expression(self, node: ASTNode.ExpressionNode, builder: ir.IRBuilder, var_type, **kwargs):
-        if node.node_type == NodeType.REFERENCE and node.value == '&': 
-            return self.handle_pointer(node, builder)
-        elif node.node_type == NodeType.BINARY_OP:
-            return self.handle_binary_expression(node, builder, var_type)
-        elif node.node_type == NodeType.LITERAL:
-            # First check if this is actually a variable reference
-            if node.value in self.symbol_table:
-                # It's a variable name, load its value
-                var_ptr = self.symbol_table[node.value]
-                return builder.load(var_ptr, name=f"load_{node.value}")
-            else:
-                # It's an actual literal value, create a constant
-                try:
-                    return ir.Constant(var_type, int(node.value))
-                except ValueError:
-                    raise ValueError(f"Invalid literal or undefined variable: '{node.value}'")
-
-    
-    def handle_expression(self, node: ASTNode.ExpressionNode, builder: ir.IRBuilder, var_type, **kwargs):
-
-        pointer_level = kwargs.get('pointer_level', 0)
-        
-        if node is None:
-            raise ValueError("Node is None, cannot handle expression.")
-
-    
-        match node.node_type:
-            case NodeType.BINARY_OP:
-                return self.handle_binary_expression(node, builder, var_type)
-
-            case NodeType.REFERENCE:
-                return self.handle_pointer(node, builder)
-
-            case NodeType.UNARY_OP:
-                return self.handle_unary_op(node, builder, var_type)
-
-            case NodeType.FUNCTION_CALL:
-                return self.handle_function_call(node, builder, **kwargs)
-
-            case NodeType.STRUCT_ACCESS:
-                return self.handle_struct_access(node, builder)
-
-            case NodeType.LITERAL:
-                return self._expression_handle_literal(node, builder, var_type, pointer_level=pointer_level)
-
-            case NodeType.REFERENCE:
-                return self.handle_pointer(node, builder)
-            case NodeType.ENUM_ACCESS:
-                return self.handle_enum_access(node, builder, var_type)
-
-            case _:
-                raise ValueError(f"Unsupported expression node type: {node.node_type}")
-    def handle_enum_access(self, node: ASTNode.ExpressionNode, builder: ir.IRBuilder, var_type, **kwargs):
-        """
-        Handle enum access operations like EnumName::MemberName or EnumName.MemberName
-        
-        Args:
-            node: The enum access node (should have enum_name and member_name attributes)
-            builder: The LLVM IR builder
-            var_type: The expected type of the expression
-            
-        Returns:
-            The LLVM constant value representing the enum member
-        """
-        debug = getattr(self.compiler, 'debug', False)
-        
-        # Extract enum name and member name from the node
-        # The exact attribute names depend on how your AST represents enum access
-        if hasattr(node, 'enum_name') and hasattr(node, 'member_name'):
-            enum_name = node.enum_name
-            member_name = node.member_name
-        elif hasattr(node, 'left') and hasattr(node, 'right'):
-            # Alternative structure where left is enum name, right is member name
-            enum_name = node.left.value if hasattr(node.left, 'value') else str(node.left)
-            member_name = node.right.value if hasattr(node.right, 'value') else str(node.right)
-        else:
-            raise ValueError(f"Invalid enum access node structure: {node}")
-        
-        if debug:
-            print(f"Handling enum access: {enum_name}::{member_name}")
-        
-        # Look up the enum type information
-        enum_type_info = Datatypes.get_type(enum_name)
-        if not enum_type_info:
-            raise ValueError(f"Unknown enum type: {enum_name}")
-        
-        # Check if it's actually an enum type
-        if not isinstance(enum_type_info, self.EnumTypeInfo):
-            raise ValueError(f"'{enum_name}' is not an enum type")
-        
-        # Get the member value
-        member_value = enum_type_info.get_member_value(member_name)
-        if member_value is None:
-            available_members = list(enum_type_info.values.keys())
-            raise ValueError(f"Enum member '{member_name}' not found in enum '{enum_name}'. "
-                            f"Available members: {available_members}")
-        
-        if debug:
-            print(f"Enum {enum_name}::{member_name} has value {member_value}")
-        
-        # Determine the target type for the constant
-        target_type = var_type if var_type else enum_type_info.get_llvm_type()
-        
-        # Create and return the constant value
-        return ir.Constant(target_type, member_value)
-
-
-    def handle_unary_op(self, node: ASTNode.ExpressionNode, builder: ir.IRBuilder, var_type, **kwargs):
-        """
-        Handle unary operations like negation (-), bitwise not (~), logical not (!), etc.
-        
-        Args:
-            node: The unary operation node
-            builder: The LLVM IR builder
-            var_type: The expected type of the expression
-            
-        Returns:
-            The LLVM value representing the result of the unary operation
-        """
-        # Process the operand first
-        operand = self.handle_expression(node.left, builder, var_type)
-        
-        match node.op:
-            case '-':  # Numeric negation
-                if isinstance(var_type, ir.FloatType):
-                    return builder.fneg(operand, name="neg")
-                else:
-                    # For integers, we can use 0 - value
-                    zero = ir.Constant(operand.type, 0)
-                    return builder.sub(zero, operand, name="neg")
-            
-            case '~':  # Bitwise NOT
-                # Only applicable to integer types
-                if isinstance(var_type, (ir.IntType)):
-                    return builder.not_(operand, name="bitnot")
-                else:
-                    raise ValueError(f"Bitwise NOT (~) cannot be applied to type {var_type}")
-            
-            case '!':  # Logical NOT
-                # Convert to boolean (0 or 1) first if not already
-                if operand.type != ir.IntType(1):
-                    # For integers, compare with 0
-                    if isinstance(operand.type, ir.IntType):
-                        zero = ir.Constant(operand.type, 0)
-                        bool_val = builder.icmp_ne(operand, zero, name="to_bool")
-                    # For floats, compare with 0.0
-                    elif isinstance(operand.type, ir.FloatType):
-                        zero = ir.Constant(operand.type, 0.0)
-                        bool_val = builder.fcmp_one(operand, zero, name="to_bool")
-                    else:
-                        raise ValueError(f"Cannot convert type {operand.type} to boolean")
-                else:
-                    bool_val = operand
-                    
-                # Invert the boolean value
-                return builder.not_(bool_val, name="lognot")
-                
-            case '*':  # Dereference pointer
-                # Check if operand is a pointer type
-                if not isinstance(operand.type, ir.PointerType):
-                    raise ValueError(f"Cannot dereference non-pointer type {operand.type}")
-                return builder.load(operand, name="deref")
-                
-            case '&':  # Reference/Address-of
-                # This should be handled elsewhere since we likely need the variable name
-                # and not just the expression result
-                raise ValueError("Address-of operator (&) should be handled by handle_pointer method")
-                
-            case _:
-                raise ValueError(f"Unsupported unary operator: {node.operator}")
-
-
-    def handle_pointer(self, node: ASTNode.ExpressionNode, builder: ir.IRBuilder, **kwargs):
-        var_ptr = self.get_variable_pointer(node.left.value)
-        return var_ptr
-
-    def _process_escape_sequences(self, string_content: str) -> str:
-        """
-        Process escape sequences in string literals.
-        
-        Args:
-            string_content: The string content without quotes
-            
-        Returns:
-            String with escape sequences processed
-        """
-        # Handle common escape sequences
-        escape_map = {
-            '\\n': '\n',
-            '\\t': '\t',
-            '\\r': '\r',
-            '\\\\': '\\',
-            '\\"': '"',
-            "\\'": "'",
-            '\\0': '\0'
-        }
-        
-        result = string_content
-        for escape_seq, replacement in escape_map.items():
-            result = result.replace(escape_seq, replacement)
-        
-        return result
-
-
-    def _create_string_literal(self, quoted_string: str, builder: ir.IRBuilder, target_type, is_pointer: bool = False) -> ir.Value:
-        """
-        Create a string literal in LLVM IR with proper PIC support.
-        
-        Args:
-            quoted_string: The string including quotes (e.g., '"Hello"')
-            builder: The LLVM IR builder
-            target_type: The target type (should be pointer type)
-            
-        Returns:
-            LLVM value representing the string literal
-        """
-        # Remove the quotes
-        string_content = quoted_string[1:-1]
-        
-        # Handle escape sequences
-        string_content = self._process_escape_sequences(string_content)
-        
-        # Special handling for single byte types (U8/I8)
-        if isinstance(target_type, ir.IntType) and target_type.width == 8 and is_pointer is False:
-            if len(string_content) == 0:
-                # Empty string -> null character
-                return ir.Constant(target_type, 0)
-            elif len(string_content) == 1:
-                # Single character -> return its ASCII value
-                return ir.Constant(target_type, ord(string_content[0]))
-            else:
-                # Multi-character string -> take first character with warning
-                print(f"Warning: String literal '{quoted_string}' truncated to first character for U8 variable")
-                return ir.Constant(target_type, ord(string_content[0]))
-        
-        # For pointer types, create a proper string with PIC support
-        # Add null terminator
-        string_with_null = string_content + '\0'
-        
-        # Create the array type for the string
-        string_array_type = ir.ArrayType(ir.IntType(8), len(string_with_null))
-        
-        # Create a global variable for the string with proper linkage for PIC
-        string_global = ir.GlobalVariable(
-            self.module, 
-            string_array_type, 
-            name=f"str_{len(getattr(self, 'string_literals', []))}"
-        )
-        
-        # Set proper attributes for PIC compilation
-        string_global.linkage = 'private'  # Changed from 'internal' to 'private'
-        string_global.global_constant = True
-        string_global.unnamed_addr = True  # Allow merging of identical constants
-        
-        # Initialize with the string content
-        string_global.initializer = ir.Constant(
-            string_array_type, 
-            bytearray(string_with_null.encode('utf-8'))
-        )
-        
-        # Keep track of string literals
-        if not hasattr(self, 'string_literals'):
-            self.string_literals = []
-        self.string_literals.append(string_global)
-        
-        # Create a GEP instruction to get pointer to the first character
-        zero = ir.Constant(ir.IntType(32), 0)
-        string_ptr = builder.gep(string_global, [zero, zero], name="str_ptr")
-        
-        return string_ptr
-
-    # Alternative approach - if you want to be more strict about type checking
-    def _create_string_literal_strict(self, quoted_string: str, builder: ir.IRBuilder, target_type):
-        """
-        Strict version that provides better error messages for type mismatches.
-        """
-        # Remove the quotes
-        string_content = quoted_string[1:-1]
-        
-        # Handle escape sequences
-        string_content = self._process_escape_sequences(string_content)
-        
-        # Strict type checking for single byte types
-        if isinstance(target_type, ir.IntType) and target_type.width == 8:
-            if len(string_content) == 0:
-                return ir.Constant(target_type, 0)  # Empty string -> null char
-            elif len(string_content) == 1:
-                return ir.Constant(target_type, ord(string_content[0]))  # Single char
-            else:
-                # Provide helpful error message
-                raise TypeError(f"Cannot assign string literal '{quoted_string}' to U8 variable. "
-                            f"String has {len(string_content)} characters but U8 can only hold 1. "
-                            f"Consider:\n"
-                            f"  - Using a single character: '\"T\"' (first char of '{string_content}')\n"
-                            f"  - Declaring variable as string pointer: 'var a: *U8 = {quoted_string}'\n"
-                            f"  - Using an array: 'var a: [4]U8 = ...' for fixed-size strings")
-    
-            
-    def _expression_handle_literal(self, node: ASTNode.ExpressionNode, builder: ir.IRBuilder, var_type, **kwargs):
-            pointer_level = kwargs.get('pointer_level', 0)
-            
-            if self.compiler.debug:
-                print(f"Handling literal: '{node.value}', target type: {var_type}, pointer_level: {pointer_level}")
-            
-            """
-            Handle literal expressions like numbers, booleans, strings, etc.
-            """
-            # Print debug info
-            debug = getattr(self, 'debug', False)
-            if debug:
-                print(f"Handling literal: '{node.value}', target type: {var_type}, pointer_level: {pointer_level}")
-            
-            # Handle special node types that might be misidentified as literals
-            if not hasattr(node, 'value'):
-                if debug:
-                    print(f"Node doesn't have a 'value' attribute: {node}")
-                if hasattr(node, 'node_type') and node.node_type == NodeType.STRUCT_ACCESS:
-                    return self.handle_struct_access(node, builder)
-                else:
-                    raise ValueError(f"Invalid literal node: {node}")
-            
-            # Check for enum access (EnumName::MemberName format in string)
-            if isinstance(node.value, str) and '::' in node.value:
-                parts = node.value.split('::', 1)
-                if len(parts) == 2:
-                    enum_name, member_name = parts
-                    enum_type_info = Datatypes.get_type(enum_name)
-                    if enum_type_info and isinstance(enum_type_info, self.EnumTypeInfo):
-                        member_value = enum_type_info.get_member_value(member_name)
-                        if member_value is not None:
-                            target_type = var_type if var_type else enum_type_info.get_llvm_type()
-                            return ir.Constant(target_type, member_value)
-                        else:
-                            available_members = list(enum_type_info.values.keys())
-                            raise ValueError(f"Enum member '{member_name}' not found in enum '{enum_name}'. "
-                                        f"Available members: {available_members}")
-            
-            # Handle variable reference if it's in the symbol table
-            if isinstance(node.value, str) and node.value in self.symbol_table:
-                var_info = self.symbol_table.lookup(node.value)
-                if var_info:
-                    # For pointer assignments, we might need to handle dereferencing
-                    if pointer_level > 0 and var_info.pointer_level > pointer_level:
-                        # Need to dereference the variable
-                        loaded_value = builder.load(var_info.llvm_value, name=f"load_{node.value}")
-                        # Apply additional dereferencing if needed
-                        for _ in range(var_info.pointer_level - pointer_level):
-                            loaded_value = builder.load(loaded_value, name=f"deref_{node.value}")
-                        return loaded_value
-                    elif pointer_level > 0 and var_info.pointer_level == 0:
-                        # Taking address of a regular variable
-                        return var_info.llvm_value  # Return the alloca (address)
-                    else:
-                        # Regular load
-                        return builder.load(var_info.llvm_value, name=f"load_{node.value}")
-            
-            # Check if this is a string literal (quoted string)
-            is_string_literal = (isinstance(node.value, str) and 
-                                len(node.value) >= 2 and 
-                                node.value.startswith('"') and 
-                                node.value.endswith('"'))
-            
-            # Infer type if not specified
-            if var_type is None:
-                if isinstance(node.value, str):
-                    if is_string_literal:
-                        # String literal - return pointer to i8 array
-                        var_type = ir.PointerType(ir.IntType(8))
-                    elif node.value.isdigit():
-                        var_type = ir.IntType(32)  # default to i32
-                    elif node.value.lower() in ['true', 'false']:
-                        var_type = ir.IntType(1)
-                    elif '.' in node.value and all(c.isdigit() or c == '.' or c == '-' or c == '+' or c.lower() == 'e' 
-                                                for c in node.value):
-                        var_type = ir.DoubleType()  # or FloatType
-                    else:
-                        # Could be a variable name or other identifier
-                        raise ValueError(f"Cannot infer type for literal value: '{node.value}'")
-                else:
-                    # If not a string, what is it?
-                    raise ValueError(f"Unsupported literal type: {type(node.value)}")
-
-            try:
-                # Handle string literals
-                if is_string_literal:
-                    # String literals are inherently pointers to char arrays
-                    return self._create_string_literal(node.value, builder, var_type, pointer_level)
-                
-                # Boolean literals (true/false)
-                if isinstance(var_type, ir.IntType) and var_type.width == 1:
-                    if isinstance(node.value, str) and node.value.lower() == 'true':
-                        return ir.Constant(var_type, 1)
-                    elif isinstance(node.value, str) and node.value.lower() == 'false':
-                        return ir.Constant(var_type, 0)
-                    elif isinstance(node.value, (int, float)):
-                        # Convert numeric value to boolean (0 = false, non-zero = true)
-                        return ir.Constant(var_type, 1 if node.value != 0 else 0)
-
-                # Handle NULL pointer literals
-                if isinstance(var_type, ir.PointerType) and (
-                    (isinstance(node.value, str) and (node.value == "0" or node.value.upper() == "NULL")) or
-                    (isinstance(node.value, (int, float)) and node.value == 0)
-                ):
-                    return ir.Constant(var_type, None)
-
-                # Handle multilevel NULL pointers
-                if pointer_level > 0 and (
-                    (isinstance(node.value, str) and (node.value == "0" or node.value.upper() == "NULL")) or
-                    (isinstance(node.value, (int, float)) and node.value == 0)
-                ):
-                    # Create appropriate null pointer type based on pointer level
-                    null_type = var_type
-                    for _ in range(pointer_level):
-                        null_type = ir.PointerType(null_type)
-                    return ir.Constant(null_type, None)
-
-                # Integer literals
-                if isinstance(var_type, ir.IntType):
-                    # Handle different types of input for integer literals
-                    if hasattr(self, '_expression_parse_integer_literal'):
-                        return self._expression_parse_integer_literal(node.value, var_type)
-                    else:
-                        # Fallback if the helper method doesn't exist
-                        if isinstance(node.value, str):
-                            # Handle hexadecimal, octal, binary literals
-                            if node.value.startswith('0x') or node.value.startswith('0X'):
-                                int_val = int(node.value, 16)
-                            elif node.value.startswith('0b') or node.value.startswith('0B'):
-                                int_val = int(node.value, 2)
-                            elif node.value.startswith('0') and len(node.value) > 1 and node.value[1].isdigit():
-                                int_val = int(node.value, 8)
-                            else:
-                                # Try parsing as decimal
-                                int_val = int(float(node.value))
-                        else:
-                            # Already a numeric value
-                            int_val = int(node.value)
-                        return ir.Constant(var_type, int_val)
-
-                # Floating point literals
-                if isinstance(var_type, (ir.FloatType, ir.DoubleType)):
-                    if isinstance(node.value, str):
-                        float_val = float(node.value)
-                    else:
-                        float_val = float(node.value)
-                    return ir.Constant(var_type, float_val)
-
-                raise ValueError(f"Unsupported literal type for value: '{node.value}' with pointer_level: {pointer_level}")
-            except ValueError as e:
-                if debug:
-                    print(f"Error handling literal: {e}")
-                raise ValueError(f"Invalid literal or undefined variable: '{node.value}'")
-        
-    def _expression_parse_integer_literal(self, value: str, var_type: ir.IntType):
-        val = int(value)
-
-        if var_type in self.signed_int_types:
-            return ir.Constant(var_type, val)
-
-        # For unsigned integers, convert negatives to 2's complement
-        if val < 0:
-            val = (1 << var_type.width) + val
-
-        return ir.Constant(var_type, val)
-
-    def handle_struct_access(self, node: ASTNode.ExpressionNode, builder: ir.IRBuilder):
-        debug = self.compiler.debug
-        """
-        Handle struct access operations, including nested access chains like a.b.c
-        This implementation flattens the nested access into a sequence of single accesses
-        """
-        # Flatten the chain of struct accesses
-        access_chain = self._flatten_struct_access(node)
-        if debug:
-            print(f"Access chain: {access_chain}")
-        
-        # Start with the base struct
-        base_name = access_chain[0]
-        base_info = self.symbol_table.lookup(base_name)
-        if not base_info:
-            raise ValueError(f"Unknown struct variable: {base_name}")
-        
-        current_ptr = base_info.llvm_value
-        current_type = base_info.data_type
-        if debug:
-            print(f"Starting with base: {base_name} of type {current_type}")
-        
-        # Process each field access in the chain (except the first which is the base)
-        result_ptr = current_ptr  # Store the final result
-        
-        for i, field_name in enumerate(access_chain[1:]):
-            if debug:
-                print(f"Accessing field: {field_name} in type: {current_type}")
-            
-            # Check if current type is a valid struct
-            class_info = self.struct_table.get(current_type)
-            if not class_info:
-                raise ValueError(f"Unknown struct type: {current_type}")
-            
-            class_type = class_info.get("class_type_info")
-            if not class_type or field_name not in class_type.field_names:
-                raise ValueError(f"Field '{field_name}' not found in struct '{current_type}'.")
-            
-            field_index = class_type.field_names.index(field_name)
-            if debug:
-                print(f"Field index: {field_index}")
-
-            
-            # Get the field pointer
-            if i == 0:  # First field access (base.field)
-                field_ptr = self.get_struct_field_ptr(base_name, field_name, builder)
-            else:
-                # For nested accesses, use GEP on the current_ptr
-                field_ptr = builder.gep(current_ptr, [ir.Constant(ir.IntType(32), 0), 
-                                                    ir.Constant(ir.IntType(32), field_index)],
-                                    name=f"field_{field_name}_ptr")
-            
-            # Load the field value
-            current_ptr = builder.load(field_ptr, name=f"access_{field_name}")
-            result_ptr = current_ptr  # Update the result pointer
-            
-            # Update current type to the field's type for next iteration
-            if field_name == 'next' and current_type == 'Node':
-                # Special case for linked list - 'next' points to another Node
-                current_type = 'Node'
-            else:
-                # For other cases, you'd need a more general mechanism to determine field types
-                # This would rely on having field type information in your struct definitions
-                if debug:
-                    print(f"Need to determine type of field {field_name} in {current_type}")
-                # Default fallback - if we're at the end of the chain, we don't need the next type
-                if i == len(access_chain[1:]) - 1:
-                    break
-                else:
-                    # For intermediate accesses, must determine next type or error
-                    raise ValueError(f"Cannot determine type of field {field_name} in {current_type}")
-        
-        return result_ptr
-
-    def _flatten_struct_access(self, node):
-        """
-        Convert a nested struct access tree into a flat list of field names
-        For example, a.b.c becomes ['a', 'b', 'c']
-        """
-        if node.node_type != NodeType.STRUCT_ACCESS:
-            # If it's just a variable reference
-            return [node.value]
-        
-        # If it's a struct access node
-        left_parts = self._flatten_struct_access(node.left)
-        right_part = node.right.value
-        return left_parts + [right_part]
-
-    def handle_block(self, node: ASTNode.Block, builder: ir.IRBuilder, **kwargs):
-        """Handle a block of statements with proper scope management."""
-        # Enter a new scope for this block
-        self.symbol_table.enter_scope()
-        
-        # Process each statement in the block
-        for stmt in node.nodes:
-            self.process_node(stmt, builder=builder, **kwargs)
-        
-        # Exit the scope when done with the block
-        self.symbol_table.exit_scope()
-
-    def handle_variable_declaration(self, node: ASTNode.VariableDeclaration, builder: ir.IRBuilder, **kwargs):
-        """Handle variable declaration with multilevel pointer support."""
-        
-        # Parse the type string to determine base type and pointer level
-        base_type_name, pointer_level = count_pointer_level(node.var_type)
-        
-        # Also check the node's pointer-related attributes if they exist
-        if hasattr(node, 'is_pointer') and node.is_pointer and pointer_level == 0:
-            pointer_level = 1  # Backward compatibility
-        elif hasattr(node, 'pointer_level'):
-            pointer_level = max(pointer_level, node.pointer_level)
-        
-        if self.compiler.debug:
-            print(f"Variable declaration: {node.name}, base_type: {base_type_name}, pointer_level: {pointer_level}")
-        
-        # Get the base LLVM type
-        base_type = Datatypes.to_llvm_type(base_type_name)
-        if base_type is None:
-            raise ValueError(f"Unknown type: {base_type_name}")
-        
-        # Apply pointer levels to get the final variable type
-        var_type = apply_pointer_level(base_type, pointer_level)
-            
-       
-        # Allocate space for the variable (this creates a pointer to var_type)
-        var = builder.alloca(var_type, name=node.name)
-        
-        # Create and store the symbol in our table
-        symbol = create_variable_symbol(
-            name=node.name,
-            ast_node=node,
-            data_type=node.var_type,  # Keep original type string
-            llvm_type=var_type,
-            llvm_value=var,
-            pointer_level=pointer_level
-        )
-        self.symbol_table.define(symbol)
-        
-        # Handle initial value if present
-        if node.value:
-            if self.compiler.debug:
-                print(f"Processing initial value for {node.name}, pointer_level: {pointer_level}")
-            
-            # Pass pointer level information to expression handler
-            value = self.handle_expression(node.value, builder, base_type, pointer_level=pointer_level)
-                
-            if not value:
-                if pointer_level > 0:
-                    # Initialize to null pointer
-                    value = ir.Constant(var_type, None)
-                else:
-                    # Default to zero for non-pointer types
-                    value = ir.Constant(var_type, 0)
-            else:
-                # Apply proper type casting before storing
-                value = self._cast_value(value, var_type, builder)
-                    
-            builder.store(value, var)
-        else:
-            # Initialize with default values
-            if pointer_level > 0:
-                # Initialize pointers to null by default
-                null_ptr = ir.Constant(var_type, None)
-                builder.store(null_ptr, var)
-            # Non-pointer types can be left uninitialized or initialized to zero
-            # depending on language semantics
-
-        print(self.symbol_table.get_all_symbols())
-        
-        return var  # Return the variable pointer
-
-
-    def _cast_value_with_pointer_level(self, value: ir.Value, target_type: Any, builder: ir.IRBuilder, 
-                                    source_pointer_level: int = 0, target_pointer_level: int = 0) -> ir.Value:
-        """
-        Cast a value to target type considering pointer levels.
-        
-        Args:
-            value: The LLVM value to cast
-            target_type: The target LLVM type
-            builder: LLVM IR builder
-            source_pointer_level: Pointer level of the source value
-            target_pointer_level: Pointer level of the target type
-        
-        Returns:
-            Casted LLVM value
-        """
-        # If pointer levels match and types are compatible, return as-is
-        if source_pointer_level == target_pointer_level and value.type == target_type:
-            return value
-        
-        # Handle pointer level mismatches
-        if source_pointer_level > target_pointer_level:
-            # Need to dereference (load from pointer)
-            result = value
-            for _ in range(source_pointer_level - target_pointer_level):
-                result = builder.load(result, name="auto_deref")
-            return result
-        elif source_pointer_level < target_pointer_level:
-            # This case is more complex and might need address-of operations
-            # For now, handle simple cases
-            if source_pointer_level == 0 and target_pointer_level == 1:
-                # Taking address of a value - this typically requires the value to be in memory
-                # This might need special handling depending on the context
-                pass
-        
-        # Fall back to regular casting
-        return self._cast_value(value, target_type, builder)
-
-
-    def _get_variable_address(self, var_name: str, builder: ir.IRBuilder, required_pointer_level: int = 1) -> ir.Value:
-        """
-        Get the address of a variable with appropriate pointer level.
-        
-        Args:
-            var_name: Name of the variable
-            builder: LLVM IR builder
-            required_pointer_level: Required pointer level for the result
-        
-        Returns:
-            LLVM value representing the address with correct pointer level
-        """
-        symbol = self.symbol_table.lookup(var_name)
-        if not symbol:
-            raise ValueError(f"Variable '{var_name}' not found")
-        
-        # The symbol's llvm_value is always an alloca (pointer to the variable's type)
-        base_address = symbol.llvm_value
-        
-        if symbol.pointer_level == 0:
-            # Regular variable - return its address
-            if required_pointer_level == 1:
-                return base_address  # Address of the variable
-            else:
-                # Need multiple levels of indirection - this is complex
-                raise ValueError(f"Cannot create {required_pointer_level}-level pointer to non-pointer variable")
-        else:
-            # Pointer variable
-            if required_pointer_level == symbol.pointer_level + 1:
-                return base_address  # Address of the pointer variable
-            elif required_pointer_level == symbol.pointer_level:
-                return builder.load(base_address, name=f"load_{var_name}")  # Value of the pointer
-            elif required_pointer_level < symbol.pointer_level:
-                # Need to dereference
-                result = builder.load(base_address, name=f"load_{var_name}")
-                for _ in range(symbol.pointer_level - required_pointer_level):
-                    result = builder.load(result, name=f"deref_{var_name}")
-                return result
-            else:
-                # Need more levels - complex case
-                raise ValueError(f"Cannot create {required_pointer_level}-level pointer from {symbol.pointer_level}-level pointer")
-
-
-    def _handle_pointer_arithmetic(self, left_val: ir.Value, operator: str, right_val: ir.Value, 
-                                builder: ir.IRBuilder, left_pointer_level: int = 0, 
-                                right_pointer_level: int = 0) -> ir.Value:
-        """
-        Handle arithmetic operations involving pointers.
-        
-        Args:
-            left_val: Left operand LLVM value
-            operator: Arithmetic operator ('+', '-', etc.)
-            right_val: Right operand LLVM value
-            builder: LLVM IR builder
-            left_pointer_level: Pointer level of left operand
-            right_pointer_level: Pointer level of right operand
-        
-        Returns:
-            Result of pointer arithmetic
-        """
-        # Pointer + integer or integer + pointer
-        if (left_pointer_level > 0 and right_pointer_level == 0) or \
-        (left_pointer_level == 0 and right_pointer_level > 0):
-            
-            if operator == '+':
-                # Use GEP (GetElementPtr) for pointer arithmetic
-                if left_pointer_level > 0:
-                    return builder.gep(left_val, [right_val], name="ptr_add")
-                else:
-                    return builder.gep(right_val, [left_val], name="ptr_add")
-            elif operator == '-' and left_pointer_level > 0:
-                # Subtract integer from pointer
-                neg_right = builder.neg(right_val, name="neg_offset")
-                return builder.gep(left_val, [neg_right], name="ptr_sub")
-        
-        # Pointer - pointer (both operands are pointers)
-        elif left_pointer_level > 0 and right_pointer_level > 0 and operator == '-':
-            # Convert pointers to integers, subtract, then divide by element size
-            left_int = builder.ptrtoint(left_val, ir.IntType(64), name="ptr_to_int_left")
-            right_int = builder.ptrtoint(right_val, ir.IntType(64), name="ptr_to_int_right")
-            diff = builder.sub(left_int, right_int, name="ptr_diff")
-            
-            # Get element size (this is simplified - real implementation needs type info)
-            element_size = ir.Constant(ir.IntType(64), 1)  # Assuming byte-sized elements
-            return builder.sdiv(diff, element_size, name="ptr_distance")
-        
-        # Regular arithmetic for non-pointer cases
-        return None  # Let caller handle regular arithmetic
-
-    def get_struct_field_ptr(self, struct_name: str, field_name: str, builder: ir.IRBuilder):
-        """
-        Returns the pointer to a struct field using GEP.
-        """
-        # Ensure the struct variable exists
-        if struct_name not in self.symbol_table:
-            raise ValueError(f"Struct variable '{struct_name}' not found.")
-
-        
-        struct_info = self.symbol_table.lookup(struct_name)
-        struct_ptr = struct_info.llvm_value
-        struct_type_name = struct_info.data_type
-
-        struct_type_info = self.struct_table[struct_type_name]["class_type_info"]
-
-        if field_name not in struct_type_info.field_names:
-            raise ValueError(f"Field '{field_name}' not found in struct '{struct_type_name}'.")
-
-        field_index = struct_type_info.field_names.index(field_name)
-        zero = ir.Constant(ir.IntType(32), 0)
-        field_idx = ir.Constant(ir.IntType(32), field_index)
-
-        return builder.gep(struct_ptr, [zero, field_idx], name=f"{struct_name}_{field_name}_ptr")
-
-
-    def handle_variable_assignment(self, node: ASTNode.VariableAssignment, builder: ir.IRBuilder, **kwargs):
-        """Handle variable assignment with the new symbol table."""
-        # Get variable name and check if it's a struct field access
-        var_name = node.name
-        
-        # Check if this is a struct field assignment (contains a dot)
-        if '.' in var_name:
-            struct_name, field_name = var_name.split('.')
-            
-            # Look up the struct in the symbol table
-            struct_symbol = self.symbol_table.lookup(struct_name)
-            if not struct_symbol:
-                raise ValueError(f"Struct variable '{struct_name}' not found in symbol table.")
-            
-            # Get struct information
-            struct_ptr = struct_symbol.llvm_value
-            struct_type_name = struct_symbol.data_type
-            
-            # Ensure the struct type exists in the struct table
-            if struct_type_name not in self.struct_table:
-                raise ValueError(f"Struct type '{struct_type_name}' not found in struct table.")
-            
-            # Get the struct type definition
-            struct_type_info = self.struct_table[struct_type_name]["class_type_info"]
-            
-            # Find the field index in the struct
-            if field_name not in struct_type_info.field_names:
-                raise ValueError(f"Field '{field_name}' not found in struct '{struct_type_name}'.")
-            
-            # Get pointer to the field
-            field_ptr = self.get_struct_field_ptr(struct_name, field_name, builder)
-            
-            # Get the field type from the struct type
-            field_index = struct_type_info.field_names.index(field_name)
-            field_type = struct_type_info.llvm_type.elements[field_index]
-            
-            # Evaluate right-hand side expression
-            value = self.handle_expression(node.value, builder, field_type)
-            
-            # Handle type casting if needed
-            if value.type != field_type:
-                value = self._cast_value(value, field_type, builder)
-            
-            # Store the value in the field
-            builder.store(value, field_ptr)
-        else:
-            # Regular variable assignment
-            symbol = self.symbol_table.lookup(var_name)
-            if not symbol:
-                raise ValueError(f"Variable '{var_name}' not found in symbol table. It must be declared before assignment.")
-            
-            # Get variable pointer and type
-            var_ptr = symbol.llvm_value
-            var_type = var_ptr.type.pointee
-            
-            # Evaluate right-hand side expression
-            value = self.handle_expression(node.value, builder, var_type)
-            
-            # Handle type casting if types don't match
-            if value.type != var_type:
-                value = self._cast_value(value, var_type, builder)
-            
-            # Store the evaluated value into the variable
-            builder.store(value, var_ptr)
-
-    def _cast_value(self, value, target_type, builder):
-        """Casts a value to the target LLVM type, inserting necessary instructions."""
-        from llvmlite import ir
-
-        src_type = value.type
-
-        # Handle Same-Type Pass-Through
-        if src_type == target_type:
-            return value
-
-        # Integer to Integer
-        if isinstance(target_type, ir.IntType) and isinstance(src_type, ir.IntType):
-            if target_type.width > src_type.width:
-                return builder.sext(value, target_type, name="sext") if Datatypes.is_signed_type(str(target_type)) else builder.zext(value, target_type, name="zext")
-            else:
-                return builder.trunc(value, target_type, name="trunc")
-
-        # Float to Float
-        if isinstance(target_type, (ir.FloatType, ir.DoubleType)) and isinstance(src_type, (ir.FloatType, ir.DoubleType)):
-            if isinstance(target_type, ir.DoubleType) and isinstance(src_type, ir.FloatType):
-                return builder.fpext(value, target_type, name="fpext")
-            elif isinstance(target_type, ir.FloatType) and isinstance(src_type, ir.DoubleType):
-                return builder.fptrunc(value, target_type, name="fptrunc")
-
-        # Int to Float
-        if isinstance(target_type, (ir.FloatType, ir.DoubleType)) and isinstance(src_type, ir.IntType):
-            return builder.sitofp(value, target_type, name="sitofp") if Datatypes.is_signed_type(str(src_type)) else builder.uitofp(value, target_type, name="uitofp")
-
-        # Float to Int
-        if isinstance(target_type, ir.IntType) and isinstance(src_type, (ir.FloatType, ir.DoubleType)):
-            return builder.fptosi(value, target_type, name="fptosi") if Datatypes.is_signed_type(target_type) else builder.fptoui(value, target_type, name="fptoui")
-
-        # Pointer to Pointer
-        if isinstance(target_type, ir.PointerType) and isinstance(src_type, ir.PointerType):
-            return builder.bitcast(value, target_type, name="ptr_cast")
-        
-        # Integer to Pointer
-        if isinstance(target_type, ir.PointerType) and isinstance(src_type, ir.IntType):
-            return builder.inttoptr(value, target_type, name="int_to_ptr")
-        
-        # Pointer to Integer
-        if isinstance(target_type, ir.IntType) and isinstance(src_type, ir.PointerType):
-            return builder.ptrtoint(value, target_type, name="ptr_to_int")
-        
-        # Boolean to Integer
-        if isinstance(target_type, ir.IntType) and isinstance(src_type, ir.IntType) and src_type.width == 1:
-            return builder.zext(value, target_type, name="bool_to_int")
-        
-        # Integer to Boolean
-        if isinstance(target_type, ir.IntType) and target_type.width == 1 and isinstance(src_type, ir.IntType):
-            # Comparing with zero to convert to boolean (0 = false, non-zero = true)
-            zero = ir.Constant(src_type, 0)
-            return builder.icmp_unsigned('!=', value, zero, name="int_to_bool")
-
-        raise TypeError(f"Incompatible types for assignment: {src_type} cannot be assigned to {target_type}")
-
-        
-
-    def handle_return(self, node: ASTNode.Return, builder: ir.IRBuilder, **kwargs):
-        # If the value is a function
-        # Get return type
-        function_return_type = builder.function.function_type.return_type
-        if node.expression:
-            return_value = self.handle_expression(node.expression, builder, function_return_type)
-            builder.ret(return_value)
-        else:
-            builder.ret_void()
-            
-    def handle_if_statement(self, node: ASTNode.IfStatement, builder: ir.IRBuilder, **kwargs):
-        # Evaluate the condition
-        condition = self.handle_expression(node.condition, builder, self.type_map[Datatypes.BOOL])
-
-        condition = self.handle_expression(node.condition, builder, self.type_map[Datatypes.BOOL])
-        if condition.type != ir.IntType(1):
-            condition = builder.trunc(condition, ir.IntType(1))
-
-
-        # Create basic blocks for the 'then', 'else', and 'merge' sections
-        then_block = builder.append_basic_block("if.then")
-        else_block = builder.append_basic_block("if.else") if node.else_body else None
-        merge_block = builder.append_basic_block("if.end")
-
-        # Branch based on the condition
-        if else_block:
-            builder.cbranch(condition, then_block, else_block)
-        else:
-            builder.cbranch(condition, then_block, merge_block)
-
-        # Generate code for the 'then' block
-        builder.position_at_end(then_block)
-        for stmt in node.if_body.nodes:
-            self.process_node(stmt, builder=builder)
-        if not builder.block.is_terminated:
-            builder.branch(merge_block)
-
-        # Generate code for the 'else' block if it exists
-        if else_block:
-            builder.position_at_end(else_block)
-            for stmt in node.else_body.nodes:
-                self.process_node(stmt, builder=builder)
-            if not builder.block.is_terminated:
-                builder.branch(merge_block)
-
-        # Position the builder at the merge block
-        builder.position_at_end(merge_block)
-
-    def handle_while_loop(self, node: ASTNode.WhileLoop, builder: ir.IRBuilder):
-        # Create basic blocks for the loop
-        loop_cond_block = builder.append_basic_block("while.cond")
-        loop_body_block = builder.append_basic_block("while.body")
-        loop_end_block = builder.append_basic_block("while.end")
-
-        # Branch to the condition block
-        builder.branch(loop_cond_block)
-
-        # Generate code for the condition block
-        builder.position_at_end(loop_cond_block)
-        condition = self.handle_expression(node.condition, builder, self.type_map[Datatypes.BOOL])
-        builder.cbranch(condition, loop_body_block, loop_end_block)
-
-        # Generate code for the body block
-        builder.position_at_end(loop_body_block)
-        for stmt in node.body.nodes:
-            self.process_node(stmt, builder=builder)
-        builder.branch(loop_cond_block)
-
-        # Position the builder at the end block
-        builder.position_at_end(loop_end_block)
-
-    def handle_for_loop(self, node: ASTNode.ForLoop, builder: ir.IRBuilder, **kwargs):
-        # Create basic blocks for the for loop
-        loop_cond_block = builder.append_basic_block("for.cond")
-        loop_body_block = builder.append_basic_block("for.body")
-        loop_update_block = builder.append_basic_block("for.update")
-        loop_end_block = builder.append_basic_block("for.end")
-
-        # Initialize the loop variable(s) (if any)
-        if node.initialization:
-            self.process_node(node.initialization, builder=builder)
-
-        # Branch to the condition block
-        builder.branch(loop_cond_block)
-        
-        # Generate code for the condition block
-        builder.position_at_end(loop_cond_block)
-        if node.condition:
-            condition = self.handle_expression(node.condition, builder, self.type_map[Datatypes.BOOL])
-            builder.cbranch(condition, loop_body_block, loop_end_block)
-        else:
-            # If there's no condition, assume an infinite loop unless we break somewhere
-            builder.branch(loop_body_block)
-
-        # Generate code for the body block
-        builder.position_at_end(loop_body_block)
-        for stmt in node.body.nodes:
-            self.process_node(stmt, builder=builder)
-
-        # After the body, move to the update block
-        builder.branch(loop_update_block)
-
-        # Generate code for the update block (if any)
-        builder.position_at_end(loop_update_block)
-        if node.update:
-            self.process_node(node.update, builder=builder)
-
-        # Return to the condition block for the next iteration
-        builder.branch(loop_cond_block)
-        
-        # Position the builder at the end block
-        builder.position_at_end(loop_end_block)
-
-    def handle_comment(self, node: ASTNode.Comment, **kwargs):
-        if 'builder' in kwargs and kwargs['builder'] is not None:
-            builder = kwargs['builder']
-            comment_text = node.text
-            if node.is_inline:
-                comment_text = "INLINE: " + comment_text
-                
-            # Add a custom metadata node that we can convert to a comment when printing
-            comment_md = builder.module.add_metadata([ir.MetaDataString(builder.module, comment_text)])
-            
-    def handle_function_call(self, node, builder: ir.IRBuilder, var_type=None, **kwargs):
-        """Handle function call with the new symbol table."""
-        # Debug the AST node structure to help understand its format
-
-        
-        # Handle either dedicated FunctionCall nodes or ExpressionNode with FUNCTION_CALL type
-        if isinstance(node, ASTNode.FunctionCall):
-            func_name = node.name
-            arguments = node.arguments
-        elif node.node_type == NodeType.FUNCTION_CALL:
-            # Extract function name
-            if hasattr(node, 'name'):
-                func_name = node.name
-            elif hasattr(node, 'value') and node.value:
-                # Sometimes the function name is stored in value
-                func_name = node.value
-            elif hasattr(node, 'left') and node.left:
-                if hasattr(node.left, 'value'):
-                    func_name = node.left.value
-                elif hasattr(node.left, 'name'):
-                    func_name = node.left.name
-                else:
-                    raise ValueError("Invalid function call format: function name not found")
-            else:
-                raise ValueError("Invalid function call format: function name not found")
-            
-            # Extract arguments - handle the various ways they might be stored
-            arguments = []
-            
-            # Check if arguments are stored directly as a property
-            if hasattr(node, 'arguments') and node.arguments is not None:
-                arguments = node.arguments
-            # Check if arguments are stored in a right property
-            elif hasattr(node, 'right') and node.right is not None:
-                if isinstance(node.right, list):
-                    arguments = node.right
-                else:
-                    # If right is a single node but represents multiple arguments
-                    if hasattr(node.right, 'arguments') and node.right.arguments is not None:
-                        arguments = node.right.arguments
-                    else:
-                        arguments = [node.right]
-        else:
-            raise TypeError("Expected a function call node")
-        
-        # Look up the function in the symbol table
-        func_symbol = self.symbol_table.lookup(func_name)
-        
-        # For backward compatibility, also check the function map
-        func = None
-        if func_symbol:
-            func = func_symbol.llvm_value
-        elif func_name in self.function_map:
-            func = self.function_map[func_name]
-        
-        if not func:
-            raise ValueError(f"Function {func_name} not defined")
-        
-        # Get the expected argument types from the function type
-        expected_arg_types = func.function_type.args
-        
-        # Check if the number of arguments matches
-        if len(arguments) != len(expected_arg_types):
-            raise ValueError(f"Function {func_name} expects {len(expected_arg_types)} arguments, but {len(arguments)} were provided")
-        
-        # Prepare arguments for the function call
-        llvm_args = []
-        for i, arg in enumerate(arguments):
-            # Process each argument as an expression
-            llvm_arg = self.handle_expression(arg, builder, None)
-            
-            # Make sure the argument types match - cast if necessary
-            expected_type = expected_arg_types[i]
-            if llvm_arg.type != expected_type:
-                # Cast the argument to the expected type
-                if isinstance(expected_type, ir.IntType) and isinstance(llvm_arg.type, ir.IntType):
-                    # For integer types, perform bit casting if needed
-                    if expected_type.width > llvm_arg.type.width:
-                        llvm_arg = builder.zext(llvm_arg, expected_type)
-                    elif expected_type.width < llvm_arg.type.width:
-                        llvm_arg = builder.trunc(llvm_arg, expected_type)
-                # Add more type casting as needed
-            
-            llvm_args.append(llvm_arg)
-        
-        # Before calling, let's log what we're about to do for debugging
-        if self.compiler.debug:
-            print(f"Calling function {func_name} with {len(llvm_args)} arguments")
-            for i, arg in enumerate(llvm_args):
-                print(f"  Argument {i}: {arg}, Type: {arg.type}")
-        
-        # Make the function call in LLVM IR
-        if func.function_type.return_type == ir.VoidType():
-            # For void functions
-            builder.call(func, llvm_args)
-            return None
-        else:
-            # For functions that return a value
-            result = builder.call(func, llvm_args)
-            return result
-
-    class ClassTypeInfo:
-                def __init__(self, llvm_type, field_names, parent_type=None, node: ASTNode.Class = None):
-                    self.llvm_type = llvm_type
-                    self.field_names = field_names
-                    self.parent = parent_type
-                    self.node = node
-                
-                def get_llvm_type(self):
-                    return self.llvm_type
-                
-                def get_fields(self):
-                    return [(name, self.llvm_type.elements[i]) for i, name in enumerate(self.field_names)]
-                
-                def get_field_index(self, field_name):
-                    try:
-                        return self.field_names.index(field_name)
-                    except ValueError:
-                        if self.parent:
-                            # Check if the field exists in the parent class
-                            for i, (name, _) in enumerate(self.parent.get_fields()):
-                                if name == field_name:
-                                    return i
-                        raise Exception(f"Unknown field '{field_name}' in class '{self.node}'")
-                def __repr__(self):
-                    return f"<ClassTypeInfo: fields={self.field_names}, parent={self.parent}, llvm_type={self.llvm_type}>"
-
-    def handle_class(self, node: ASTNode.Class, **kwargs):
-        # Get the parent class info if any
-        parent_type_info = None
-        if node.parent:
-            parent_type_info = Datatypes.get_type(node.parent)
-            if not parent_type_info:
-                raise Exception(f"Unknown parent class '{node.parent}'")
-
-        # --- START: Handling Identified LLVM Struct Types ---
-
-        # 1. Create the identified (named) struct type in the LLVM context.
-        # This type is initially 'opaque' (its contents are not yet defined).
-        # We use a standard naming convention like '%struct.ClassName'.
-        llvm_struct_name = f"%struct.{node.name}"
-        # Use global_context to get the type by name. If it doesn't exist, it's created.
-        struct_type = ir.global_context.get_identified_type(llvm_struct_name)
-
-        # 2. Collect the LLVM types for each field in the class.
-        # We need to do this *after* creating the identified type so that
-        # self-referential fields (like 'Node next' in a Node class) can
-        # correctly refer to a pointer to 'struct_type'.
-        field_llvm_types = []
-        field_names = []
-
-        # If there's a parent, include its field types first (inheritance).
-        # Ensure inherited_fields provides LLVM types compatible with the parent's struct layout.
-        if parent_type_info and hasattr(parent_type_info, 'get_fields'):
-            inherited_fields = parent_type_info.get_fields() # Assuming this returns [(name, llvm_type)]
-            for field_name, field_type in inherited_fields:
-                field_names.append(field_name)
-                field_llvm_types.append(field_type)
-
-        # Process each field defined directly in this class.
-        for field in node.fields:
-            # Convert the source type name to its corresponding LLVM type.
-            # Special handling is needed here for fields that are pointers to *this* class.
-            if field.var_type == node.name:
-                # If the field type is the same as the class being defined,
-                # it should be a pointer to this identified struct type.
-                field_llvm_type = struct_type.as_pointer()
-            else:
-                # For other types, use the standard conversion.
-                # Datatypes.to_llvm_type should handle base types (U8, etc.)
-                # and potentially look up other defined class types (usually returning pointers).
-                field_llvm_type = Datatypes.to_llvm_type(field.var_type)
-
-            field_llvm_types.append(field_llvm_type)
-            field_names.append(field.name)
-
-        # 3. Set the body of the identified struct type.
-        # This defines the actual layout (the sequence of field types) for the struct.
-        # This step completes the definition of the 'opaque' type created earlier.
-        struct_type.set_body(*field_llvm_types) # Use * to unpack the list of types
-
-        # --- END: Handling Identified LLVM Struct Types ---
-
-        # Create a wrapper object to store additional information about our class,
-        # including the now-defined LLVM struct type.
-        class_type_info = self.ClassTypeInfo(struct_type, field_names, parent_type_info, node)
-
-        # Register the class type info in your type system.
-        # This makes the 'Node' source type name map to the 'class_type_info' object,
-        # which contains the LLVM 'struct_type'.
-        Datatypes.add_type(node.name, class_type_info)
-
-        # Store the struct info in your internal table (if needed).
-        # Ensure you are storing the class name as the key.
-        self.struct_table[node.name] = {'name': node.name, 'class_type_info': class_type_info}
-
-        # This function typically doesn't return an LLVM value, just defines the type.
-        return None
-
-
-
-    def _determine_enum_type(self, members):
-        """
-        Determine if this is a string enum or integer enum based on the first member
-        
-        Args:
-            members: List of (member_name, value_expr) tuples
-            
-        Returns:
-            EnumValueType indicating the type of enum
-        """
-        if not members:
-            return EnumValueType.INTEGER  # Default to integer for empty enums
-        
-        # Check the first member that has a value
-        for member_name, value_expr in members:
-            if value_expr is not None:
-                if value_expr.node_type == NodeType.LITERAL:
-                    # Check if the literal value is a string (contains quotes)
-                    if isinstance(value_expr.value, str):
-                        # Check if it's a quoted string
-                        if (value_expr.value.startswith('"') and value_expr.value.endswith('"')) or \
-                        (value_expr.value.startswith("'") and value_expr.value.endswith("'")):
-                            return EnumValueType.STRING
-                        else:
-                            # Try to parse as integer
-                            try:
-                                int(value_expr.value)
-                                return EnumValueType.INTEGER
-                            except ValueError:
-                                # If it's not a valid integer, treat as string
-                                return EnumValueType.STRING
-        
-        # Default to integer enum
-        return EnumValueType.INTEGER
-
-    def _create_string_constant(self, builder, string_value):
-        """
-        Create a string constant and return a pointer to it
-        
-        Args:
-            builder: LLVM IR builder
-            string_value: The string value (with or without quotes)
-            
-        Returns:
-            LLVM constant representing i8* pointer to the string
-        """
-        # Remove quotes if present
-        if string_value.startswith('"') and string_value.endswith('"'):
-            clean_string = string_value[1:-1]
-        elif string_value.startswith("'") and string_value.endswith("'"):
-            clean_string = string_value[1:-1]
-        else:
-            clean_string = string_value
-        
-        # Handle escape sequences
-        clean_string = clean_string.replace('\\n', '\n').replace('\\t', '\t').replace('\\"', '"').replace("\\'", "'")
-        
-        # Create the string constant
-        string_const = ir.Constant(ir.ArrayType(ir.IntType(8), len(clean_string) + 1), 
-                                bytearray(clean_string.encode('utf-8') + b'\0'))
-        
-        # Create a global variable to hold the string
-        global_string = ir.GlobalVariable(self.module, string_const.type, name=f"str_{hash(clean_string) & 0xFFFFFFFF}")
-        global_string.initializer = string_const
-        global_string.global_constant = True
-        global_string.linkage = 'private'
-        
-        # Return a pointer to the string (i8*)
-        return global_string.gep([ir.Constant(ir.IntType(32), 0), ir.Constant(ir.IntType(32), 0)])
-
-    def handle_enum(self, node: ASTNode.Enum, **kwargs):
-        """
-        Enhanced enum handler that supports both integer and string enums
-        """
-        debug = getattr(self.compiler, 'debug', False)
-        
-        if debug:
-            print(f"Processing enum: {node.name}")
-        
-        # Initialize enum table if it doesn't exist
-        if not hasattr(self, 'enum_table'):
-            self.enum_table = EnumTable()
-        
-        # Determine the type of enum (integer or string)
-        enum_type = self._determine_enum_type(node.members)
-        
-        if debug:
-            print(f"Enum type determined: {enum_type.value}")
-        
-        # Set up the underlying LLVM type
-        if enum_type == EnumValueType.STRING:
-            underlying_type = ir.PointerType(ir.IntType(8))  # i8* for strings
-        else:
-            underlying_type = ir.IntType(32)  # i32 for integers
-        
-        # Process enum members
-        enum_values = {}
-        llvm_constants = {}
-        current_value = 0  # For auto-incrementing integer enums
-        
-        for member_name, value_expr in node.members:
-            if debug:
-                print(f"Processing member: {member_name}")
-            
-            if enum_type == EnumValueType.STRING:
-                if value_expr is not None and value_expr.node_type == NodeType.LITERAL:
-                    string_value = value_expr.value
-                    enum_values[member_name] = string_value
-                    
-                    # Create the LLVM string constant
-                    string_constant = self._create_string_constant(self.builder, string_value)
-                    llvm_constants[member_name] = string_constant
-                    
-                    if debug:
-                        print(f"String enum member {member_name} = {string_value}")
-                else:
-                    raise ValueError(f"String enum member '{member_name}' must have a string literal value")
-            
-            else:  # Integer enum
-                if value_expr is not None:
-                    # Evaluate the constant expression
-                    evaluated_value = self._evaluate_enum_constant(value_expr)
-                    if evaluated_value is not None:
-                        current_value = evaluated_value
-                
-                enum_values[member_name] = current_value
-                if debug:
-                    print(f"Integer enum member {member_name} = {current_value}")
-                
-                current_value += 1  # Auto-increment for next member
-        
-        # Create the EnumTypeInfo object
-        enum_type_info = EnumTypeInfo(
-            name=node.name,
-            enum_value_type=enum_type,
-            llvm_type=underlying_type,
-            values=enum_values,
-            ast_node=node
-        )
-        
-        # For string enums, store the LLVM constants
-        if enum_type == EnumValueType.STRING:
-            enum_type_info.llvm_constants = llvm_constants
-        
-        # Register the enum in the type system
-        Datatypes.add_type(node.name, enum_type_info)
-        
-        # Add to our enum table
-        self.enum_table.add_enum(node.name, enum_type_info)
-        
-        if debug:
-            print(f"Enum {node.name} registered successfully")
-        
-        return None
-
-    def handle_enum_access(self, node: ASTNode.ExpressionNode, builder: ir.IRBuilder, var_type, **kwargs):
-        """
-        Enhanced enum access handler that supports both integer and string enums
-        """
-        debug = getattr(self.compiler, 'debug', False)
-        
-        # Extract enum name and member name from the node
-        if hasattr(node, 'enum_name') and hasattr(node, 'member_name'):
-            enum_name = node.enum_name
-            member_name = node.member_name
-        elif hasattr(node, 'left') and hasattr(node, 'right'):
-            enum_name = node.left.value if hasattr(node.left, 'value') else str(node.left)
-            member_name = node.right.value if hasattr(node.right, 'value') else str(node.right)
-        else:
-            raise ValueError(f"Invalid enum access node structure: {node}")
-        
-        if debug:
-            print(f"Handling enum access: {enum_name}::{member_name}")
-        
-        # Look up the enum type information
-        if hasattr(self, 'enum_table'):
-            enum_type_info = self.enum_table.get_enum(enum_name)
-        else:
-            enum_type_info = Datatypes.get_type(enum_name)
-        
-        if not enum_type_info:
-            raise ValueError(f"Unknown enum type: {enum_name}")
-        
-        if not isinstance(enum_type_info, EnumTypeInfo):
-            raise ValueError(f"'{enum_name}' is not an enum type")
-        
-        # Check if the member exists
-        if not enum_type_info.has_member(member_name):
-            available_members = list(enum_type_info.values.keys())
-            raise ValueError(f"Enum member '{member_name}' not found in enum '{enum_name}'. "
-                            f"Available members: {available_members}")
-        
-        if debug:
-            member_value = enum_type_info.get_member_value(member_name)
-            print(f"Enum {enum_name}::{member_name} has value {member_value} (type: {enum_type_info.enum_value_type.value})")
-        
-        # Return the appropriate LLVM constant
-        if enum_type_info.is_string_enum():
-            # For string enums, return the stored LLVM constant (i8*)
-            llvm_constant = enum_type_info.get_llvm_constant(member_name)
-            if llvm_constant is None:
-                raise ValueError(f"LLVM constant not found for string enum member {enum_name}::{member_name}")
-            return llvm_constant
-        else:
-            # For integer enums, create the constant
-            member_value = enum_type_info.get_member_value(member_name)
-            target_type = var_type if var_type else enum_type_info.get_llvm_type()
-            return ir.Constant(target_type, member_value)
-
-    # Helper method additions to your existing class
-    def _evaluate_enum_constant(self, expr):
-        """
-        Enhanced constant expression evaluator (your existing implementation with minor improvements)
-        """
-        if expr.node_type == NodeType.LITERAL:
-            if isinstance(expr.value, str):
-                # Check if it's a quoted string (should not be evaluated as integer)
-                if (expr.value.startswith('"') and expr.value.endswith('"')) or \
-                (expr.value.startswith("'") and expr.value.endswith("'")):
-                    return None  # String literals are not integer constants
-                
-                try:
-                    # Handle different number formats
-                    if expr.value.startswith('0x') or expr.value.startswith('0X'):
-                        return int(expr.value, 16)
-                    elif expr.value.startswith('0b') or expr.value.startswith('0B'):
-                        return int(expr.value, 2)
-                    elif expr.value.startswith('0') and len(expr.value) > 1 and expr.value[1].isdigit():
-                        return int(expr.value, 8)
-                    else:
-                        return int(expr.value)
-                except ValueError:
-                    return None  # Not a valid integer
-            elif isinstance(expr.value, (int, float)):
-                return int(expr.value)
-        
-        elif expr.node_type == NodeType.BINARY_OP:
-            # Your existing binary operation handling
-            left_val = self._evaluate_enum_constant(expr.left)
-            right_val = self._evaluate_enum_constant(expr.right)
-            
-            if left_val is None or right_val is None:
-                return None
-                
-            match expr.op:
-                case '+':
-                    return left_val + right_val
-                case '-':
-                    return left_val - right_val
-                case '*':
-                    return left_val * right_val
-                case '/':
-                    return left_val // right_val
-                case '%':
-                    return left_val % right_val
-                case '<<':
-                    return left_val << right_val
-                case '>>':
-                    return left_val >> right_val
-                case '&':
-                    return left_val & right_val
-                case '|':
-                    return left_val | right_val
-                case '^':
-                    return left_val ^ right_val
-                case _:
-                    return None
-        
-        elif expr.node_type == NodeType.UNARY_OP:
-            # Your existing unary operation handling
-            operand_val = self._evaluate_enum_constant(expr.left)
-            if operand_val is None:
-                return None
-                
-            match expr.op:
-                case '-':
-                    return -operand_val
-                case '+':
-                    return operand_val
-                case '~':
-                    return ~operand_val
-                case _:
-                    return None
-        
-        return None 
-    def handle_union(self, node, **kwargs):
-        pass
-
-    def handle_break(self, node, **kwargs):
-        pass
-
-    def handle_continue(self, node, **kwargs):
-        pass
-
-    def handle_variable_increment(self, node, builder, **kwargs):
-        # Find variable in symbol table
-        self.symbol_table.lookup(node.name)
-        # Get the variable pointer
-        var_ptr = self.get_variable_pointer(node.name)
-        # Load the current value
-        current_value = builder.load(var_ptr, name=f"load_{node.name}")
-        # Increment the value
-        incremented_value = builder.add(current_value, ir.Constant(current_value.type, 1), name=f"increment_{node.name}")
-        # Store the incremented value back to the variable
-        builder.store(incremented_value, var_ptr)
-        # Return the incremented value
-        return incremented_value
-    
-    def handle_variable_decrement(self, node, builder, **kwargs):
-        # Find variable in symbol table
-        self.symbol_table.lookup(node.name)
-        # Get the variable pointer
-        var_ptr = self.get_variable_pointer(node.name)
-        # Load the current value
-        current_value = builder.load(var_ptr, name=f"load_{node.name}")
-        # Decrement the value
-        decremented_value = builder.sub(current_value, ir.Constant(current_value.type, 1), name=f"decrement_{node.name}")
-        # Store the decremented value back to the variable
-        builder.store(decremented_value, var_ptr)
-        # Return the decremented value
-        return decremented_value
-
-            
-    def handle_extern(self, node: ASTNode.Extern, **kwargs):
-        print(node)
-        """Handle extern function declarations."""
-        # Get function name, return type and parameter information
-        func_name = node.declaration.name
-        return_type = self.get_llvm_type(node.declaration.return_type)
-        if return_type is None:
-            raise ValueError(f"Unknown return type for function '{func_name}'")
-        
-        # Process parameters
-        param_types = []
-        for param in node.declaration.parameters:
-            param_type = self.get_llvm_type(param.var_type)
-            # Handle pointer types
-            if param.pointer_level > 0:
-                param_type = ir.PointerType(param_type)
-            param_types.append(param_type)
-        
-        # Create function type
-        func_type = ir.FunctionType(return_type, param_types, var_arg=(func_name == 'printf'))
-        
-        # Check if function already exists in the module
-        if func_name in self.module.globals:
-            # Function already declared, return the existing function
-            return self.module.globals[func_name]
-        
-        # Create the function declaration
-        func = ir.Function(self.module, func_type, name=func_name)
-        
-        # Set parameter names (optional but useful for debugging)
-        for i, param in enumerate(node.declaration.parameters):
-            func.args[i].name = param.name
-        
-        # Register the function in the symbol table
-        symbol = create_function_symbol(
-            func_name, 
-            node.declaration, 
-            node.declaration.return_type, 
-            [param.var_type for param in node.declaration.parameters],
-            func, 
-            0  # Global scope
-        )
-        self.symbol_table.define(symbol)
-        
-        # Return the function reference
-        return func
-    
-    def handle_compound_variable_declaration(self, node: ASTNode.CompoundVariableDeclaration, builder: ir.IRBuilder, **kwargs):
-        """Handle compound variable declarations (like structs or unions)."""
-        # Get the type of the compound variable
-        for declaration in node.declarations:
-            self.handle_variable_declaration(declaration, builder)
-        
-    def handle_compound_variable_assignment(self, node: ASTNode.CompoundVariableAssigment, builder: ir.IRBuilder, **kwargs):
-        for assignment in node.assignments:
-            self.handle_variable_assignment(assignment, builder)
-
